feat: floatbar
