+++ apps/renderer/src/constants/dom.ts
@@ -1,1 +1,3 @@
 export const ENTRY_CONTENT_RENDER_CONTAINER_ID = "follow-entry-render"
+
+export const LOGO_MOBILE_ID = "follow-logo-mobile"
