--- apps/renderer/src/modules/app-layout/center-column/mobile.tsx
@@ -1,9 +1,0 @@
-import { EntryColumn } from "~/modules/entry-column"
-
-export const CenterColumnMobile = () => {
-  return (
-    <div className="flex h-screen min-w-0 grow">
-      <EntryColumn />
-    </div>
-  )
-}
