+++ apps/renderer/src/modules/app-layout/entry-content/index.electron.tsx
