+++ apps/renderer/src/modules/app-layout/entry-content/index.tsx
@@ -2,11 +2,11 @@ import { withResponsiveComponent } from "@follow/components/utils/selector.js"
 
 export const RightContentLayout = withResponsiveComponent<object>(
   () =>
-    import("~/modules/app-layout/right-content/desktop").then((m) => ({
+    import("~/modules/app-layout/entry-content/desktop").then((m) => ({
       default: m.RightContentDesktop,
     })),
   () =>
-    import("~/modules/app-layout/right-content/mobile").then((m) => ({
-      default: m.RightContentMobile,
+    import("~/modules/app-layout/entry-content/mobile").then((m) => ({
+      default: m.EntryContentMobile,
     })),
 )
