+++ apps/renderer/src/modules/app-layout/entry-content/mobile.tsx
@@ -2,15 +2,15 @@ import { ROUTE_ENTRY_PENDING } from "~/constants"
 import { useRouteParamsSelector } from "~/hooks/biz/useRouteParams"
 import { EntryContent } from "~/modules/entry-content"
 
-import { CenterColumnMobile } from "../center-column/mobile"
+import { EntryColumnMobile } from "../entry-column/mobile"
 
-export const RightContentMobile = () => {
+export const EntryContentMobile = () => {
   const { entryId } = useRouteParamsSelector((s) => ({
     entryId: s.entryId,
   }))
 
   if (entryId === ROUTE_ENTRY_PENDING) {
-    return <CenterColumnMobile />
+    return <EntryColumnMobile />
   }
   return <EntryContent entryId={entryId!} />
 }
