+++ apps/renderer/src/modules/app-layout/feed-column/components/FooterInfo.tsx
