+++ apps/renderer/src/modules/app-layout/feed-column/index.mobile.tsx
@@ -1,0 +1,8 @@
+import { Outlet } from "react-router-dom"
+
+import { useDailyTask } from "~/hooks/biz/useDailyTask"
+
+export const MobileRootLayout = () => {
+  useDailyTask()
+  return <Outlet />
+}
