--- apps/renderer/src/modules/app-layout/left-sidebar/index.mobile.tsx
@@ -1,5 +1,0 @@
-import { Outlet } from "react-router-dom"
-
-export const LeftSidebarLayout = () => {
-  return <Outlet />
-}
