+++ apps/renderer/src/pages/(main)/(layer)/feeds/[feedId]/[entryId]/index.tsx
@@ -1,1 +1,1 @@
-export { RightContentLayout as Component } from "~/modules/app-layout/right-content/index"
+export { RightContentLayout as Component } from "~/modules/app-layout/entry-content/index"
