+++ apps/renderer/src/pages/(main)/layout.tsx
@@ -1,1 +1,1 @@
-export { LeftSidebarLayout as Component } from "~/modules/app-layout/left-sidebar/index"
+export { LeftSidebarLayout as Component } from "~/modules/app-layout/feed-column/index"
